# docs/swagger.yaml
...
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.DirectoryMember:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      headline:
        type: string
      open_to_collaboration:
        type: boolean
      open_to_hire:
        type: boolean
      open_to_mentoring:
        type: boolean
      pronouns:
        type: string
      reputation:
        type: integer
      tags:
        items:
          $ref: '#/definitions/schema.ProfileTagInfo'
        type: array
      timezone:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  schema.NetworkProfileUpdateReq:
    properties:
      external_links:
        items:
          $ref: '#/definitions/schema.ProfileExternalLink'
        maxItems: 10
        type: array
      headline:
        maxLength: 255
        type: string
      open_to_collaboration:
        type: boolean
      open_to_hire:
        type: boolean
      open_to_mentoring:
        type: boolean
      pronouns:
        maxLength: 64
        type: string
      timezone:
        maxLength: 64
        type: string
    type: object
  schema.NetworkProjectCreateReq:
    properties:
      description:
        maxLength: 4000
        type: string
      repo_url:
        maxLength: 512
        type: string
      seeking_help:
        type: boolean
      status:
        enum:
        - 1
        - 2
        - 9
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  schema.NetworkProjectUpdateReq:
    properties:
      description:
        maxLength: 4000
        type: string
      repo_url:
        maxLength: 512
        type: string
      seeking_help:
        type: boolean
      status:
        enum:
        - 1
        - 2
        - 9
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  schema.NetworkSetProfileTagsReq:
    properties:
      tag_ids:
        items:
          type: string
        maxItems: 30
        type: array
    required:
    - tag_ids
    type: object
...
  schema.ProfileExternalLink:
    properties:
      label:
        maxLength: 64
        type: string
      url:
        maxLength: 512
        type: string
    required:
    - label
    - url
    type: object
  schema.ProfileProjectInfo:
    properties:
      description:
        type: string
      id:
        type: string
      repo_url:
        type: string
      seeking_help:
        type: boolean
      status:
        type: integer
      title:
        type: string
      updated_at:
        type: integer
    type: object
  schema.ProfileTagInfo:
    properties:
      description:
        type: string
      id:
        type: string
      kind:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  site.NetworkProfile:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      external_links:
        items:
          $ref: '#/definitions/schema.ProfileExternalLink'
        type: array
      global_rank:
        type: integer
      headline:
        type: string
      open_to_collaboration:
        type: boolean
      open_to_hire:
        type: boolean
      open_to_mentoring:
        type: boolean
      projects:
        items:
          $ref: '#/definitions/schema.ProfileProjectInfo'
        type: array
      pronouns:
        type: string
      site_ranks:
        items:
          $ref: '#/definitions/site.NetworkProfileSiteRank'
        type: array
      tags:
        items:
          $ref: '#/definitions/schema.ProfileTagInfo'
        type: array
      timezone:
        type: string
      user_id:
        type: string
    type: object
  site.NetworkProfileSiteRank:
    properties:
      rank:
        type: integer
      site_id:
        type: string
      site_name:
        type: string
      site_slug:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/network/me/tags:
    put:
      consumes:
      - application/json
      description: replace the logged-in user's skill/interest tags
      parameters:
      - description: tags
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NetworkSetProfileTagsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: replace the caller's profile tags
      tags:
      - NetworkDirectory
  /answer/api/v1/network/members:
    get:
      description: faceted member search; members-only unless directory_visibility
        is public
      parameters:
      - description: text query
        in: query
        name: q
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.DirectoryMember'
                        type: array
                    type: object
              type: object
      summary: search the member directory
      tags:
      - NetworkDirectory
  /answer/api/v1/network/profile:
    put:
      consumes:
      - application/json
      description: update the extended directory profile of the logged-in user
      parameters:
      - description: profile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NetworkProfileUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update the caller's network profile
      tags:
      - NetworkDirectory
  /answer/api/v1/network/projects:
    post:
      consumes:
      - application/json
      description: add a project-in-progress to the logged-in user's profile
      parameters:
      - description: project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NetworkProjectCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProfileProjectInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: add a project to the caller's profile
      tags:
      - NetworkDirectory
  /answer/api/v1/network/projects/{id}:
    delete:
      description: remove a project from the logged-in user's profile
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: remove one of the caller's projects
      tags:
      - NetworkDirectory
    put:
      consumes:
      - application/json
      description: update a project on the logged-in user's profile
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      - description: project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NetworkProjectUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProfileProjectInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: update one of the caller's projects
      tags:
      - NetworkDirectory
  /answer/api/v1/network/tags:
    get:
      description: list the curated skill/interest tag catalog
      parameters:
      - description: 1=skill, 2=interest, 3=both
        in: query
        name: kind
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.ProfileTagInfo'
                  type: array
              type: object
      summary: list active profile tags
      tags:
      - NetworkDirectory
  /answer/api/v1/network/user/profile:
    get:
      description: reputation and activity across every sub-site; members-only when
        the directory is not public
      parameters:
      - description: user id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/site.NetworkProfile'
              type: object
      summary: get a user's network-wide profile
      tags:
      - Site
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"